+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Handle nullable parameters to API clients better. Only send query parameters if we have a value
+- Unpack non-builtin query parameters automatically (#68)
 
 ## [0.12.2] - 2024-09-16